テストでプロパティの参照しているところを事前に計算するように変更

+++ csharp/McapCs.Test/Record/ChannelTests.cs
@@ -38,21 +38,27 @@ public class ChannelTests
         expectedWriter.Write((byte)EOpCode.Channel);
 
         // Manually calculate metadataSize based on hardcoded strings
-        ulong metadataSize = (ulong)(
-            4 + Encoding.UTF8.GetByteCount(KEY1) + 4 + Encoding.UTF8.GetByteCount(VALUE1) +
-            4 + Encoding.UTF8.GetByteCount(KEY2) + 4 + Encoding.UTF8.GetByteCount(VALUE2)
+        ulong metadataSize = (
+            4UL + (ulong)Encoding.UTF8.GetByteCount(KEY1) + 4UL + (ulong)Encoding.UTF8.GetByteCount(VALUE1) +
+            4UL + (ulong)Encoding.UTF8.GetByteCount(KEY2) + 4UL + (ulong)Encoding.UTF8.GetByteCount(VALUE2)
         );
 
         // Calculate recordSize using hardcoded string byte counts
-        ulong recordSize = 2UL + 2UL + 4UL + (ulong)Encoding.UTF8.GetByteCount(TOPIC) + 4UL + (ulong)Encoding.UTF8.GetByteCount(MESSAGE_ENCODING) + 4UL + metadataSize;
+        ulong recordSize = (
+            2UL + // id
+            2UL + // schemaId
+            4UL + (ulong)Encoding.UTF8.GetByteCount(TOPIC) + // topic length + topic data
+            4UL + (ulong)Encoding.UTF8.GetByteCount(MESSAGE_ENCODING) + // messageEncoding length + messageEncoding data
+            4UL + metadataSize // metadata length + metadata data
+        );
         expectedWriter.Write(recordSize);
-        expectedWriter.Write((ushort)1);
-        expectedWriter.Write((ushort)2);
-        expectedWriter.Write((uint)Encoding.UTF8.GetByteCount(TOPIC));
-        expectedWriter.Write(Encoding.UTF8.GetBytes(TOPIC));
-        expectedWriter.Write((uint)Encoding.UTF8.GetByteCount(MESSAGE_ENCODING));
-        expectedWriter.Write(Encoding.UTF8.GetBytes(MESSAGE_ENCODING));
-        expectedWriter.Write((uint)metadataSize);
+        expectedWriter.Write((ushort)1); // id
+        expectedWriter.Write((ushort)2); // schemaId
+        expectedWriter.Write((uint)Encoding.UTF8.GetByteCount(TOPIC)); // topic length
+        expectedWriter.Write(Encoding.UTF8.GetBytes(TOPIC)); // topic data
+        expectedWriter.Write((uint)Encoding.UTF8.GetByteCount(MESSAGE_ENCODING)); // messageEncoding length
+        expectedWriter.Write(Encoding.UTF8.GetBytes(MESSAGE_ENCODING)); // messageEncoding data
+        expectedWriter.Write((uint)metadataSize); // metadata length
         // Manually write sorted metadata
         // Note: The order of metadata is lexicographical by key in MCAP spec
         expectedWriter.Write((uint)Encoding.UTF8.GetByteCount(KEY1));
